+++ example/main.dart
@@ -3,7 +3,7 @@
 import 'dart:io' as io;
 
 import 'package:path/path.dart' as path;
-import 'package:epub_plus/epub_plus.dart';
+import 'package:epub_pro/epub_pro.dart';
 import 'package:collection/collection.dart';
 
 void main(List<String> args) async {