+++ lib/epub_pro.dart
@@ -1,5 +1,3 @@
-library epub_plus;
-
 export 'src/utils/enum_from_string.dart';
 
 export 'src/epub_reader.dart';