+++ lib/src/writers/epub_manifest_writer.dart
@@ -1,4 +1,4 @@
-import 'package:epub_plus/src/schema/opf/epub_manifest.dart';
+import 'package:epub_pro/src/schema/opf/epub_manifest.dart';
 import 'package:xml/xml.dart' show XmlBuilder;
 
 class EpubManifestWriter {