+++ test/epub_image_tests.dart
@@ -5,7 +5,7 @@ import 'dart:io' as io;
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
-import 'package:epub_plus/epub_plus.dart';
+import 'package:epub_pro/epub_pro.dart';
 
 main() async {
   String fileName = "MY VAMPIRE SYSTEM (JKSManga) (Z-Library).epub";