+++ test/epub_writer_tests.dart
@@ -2,7 +2,7 @@ library epubreadertest;
 
 import 'dart:io' as io;
 
-import 'package:epub_plus/epub_plus.dart';
+import 'package:epub_pro/epub_pro.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';