+++ test/schema/navigation/epub_navigation_target_test.dart
@@ -2,8 +2,8 @@ library epubreadertest;
 
 import 'dart:math';
 
-import 'package:epub_plus/epub_plus.dart';
-import 'package:epub_plus/src/schema/navigation/epub_navigation_target.dart';
+import 'package:epub_pro/epub_pro.dart';
+import 'package:epub_pro/src/schema/navigation/epub_navigation_target.dart';
 import 'package:test/test.dart';
 
 import '../../random_data_generator.dart';