+++ test/schema/opf/epub_guide_test.dart
@@ -2,7 +2,7 @@ library epubreadertest;
 
 import 'dart:math' show Random;
 
-import 'package:epub_plus/src/schema/opf/epub_guide.dart';
+import 'package:epub_pro/src/schema/opf/epub_guide.dart';
 import 'package:test/test.dart';
 
 import '../../random_data_generator.dart';