+++ test/schema/opf/epub_manifest_test.dart
@@ -1,7 +1,7 @@
 library epubreadertest;
 
-import 'package:epub_plus/src/schema/opf/epub_manifest.dart';
-import 'package:epub_plus/src/schema/opf/epub_manifest_item.dart';
+import 'package:epub_pro/src/schema/opf/epub_manifest.dart';
+import 'package:epub_pro/src/schema/opf/epub_manifest_item.dart';
 import 'package:test/test.dart';
 
 main() async {