+++ test/schema/opf/epub_spine_test.dart
@@ -2,8 +2,8 @@ library epubreadertest;
 
 import 'dart:math';
 
-import 'package:epub_plus/src/schema/opf/epub_spine.dart';
-import 'package:epub_plus/src/schema/opf/epub_spine_item_ref.dart';
+import 'package:epub_pro/src/schema/opf/epub_spine.dart';
+import 'package:epub_pro/src/schema/opf/epub_spine_item_ref.dart';
 import 'package:test/test.dart';
 
 import '../../random_data_generator.dart';